Move to_float to its own file, as we need to re-sort after merging subway stations, which means importing the POIs sort function and we can't do circular imports.
