+++ TileStache/Goodies/VecTiles/sort.py
@@ -1,4 +1,4 @@
-from transform import to_float
+from util import to_float
 
 # sort functions to apply to features
