+++ TileStache/Goodies/VecTiles/transform.py
@@ -10,21 +10,8 @@ from shapely.geometry.multipoint import MultiPoint
 from shapely.geometry.multilinestring import MultiLineString
 from shapely.geometry.multipolygon import MultiPolygon
 from shapely.geometry.collection import GeometryCollection
+from util import to_float
 import re
-
-
-# attempts to convert x to a floating point value,
-# first removing some common punctuation. returns
-# None if conversion failed.
-def to_float(x):
-    if x is None:
-        return None
-    # normalize punctuation
-    x = x.replace(';', '.').replace(',', '.')
-    try:
-        return float(x)
-    except ValueError:
-        return None
 
 
 feet_pattern = re.compile('([+-]?[0-9.]+)\'(?: *([+-]?[0-9.]+)")?')
